Fix changing weapons not affecting the weapon bar
If you switched out or rearranged weapons in the ship UI, these changes
wouldn't be immediately reflected in the weapon targeting panel. This
patch adds a callback to the currently-open window to make one system
for alerting everything that anything about the ship has changed, and
all the buttons should be rebuilt.