Use real buttons for the system power buttons
The old system needed to calculate the X and Y for each system power
button independently, which was already a mess with drones and
implementing more different-width systems (such as cloaking, teleporter,
etc) would be a huge mess.

By using a button we calculate the X position once in the render loop,
where we can just accumulate the widths of all the systems which makes
implementing any new system a lot easier.