Implement ion damage
As far as I can tell, this is pretty much everything related to ion
damage: the graphics match, you can't change system power, weapons apply
it (tested with both ion blasts from enemy ships and ion bombs from the
player ship), and it reduces the system's power.

One of the motivations for implementing this is the cooldown on
cloaking, which will be able to re-use a large part of it.