Fix the DialogueWindow choice positions
This reworks everything using more in-game measurements, so now pretty
much everything should be pixel-perfect (apart from the not-in-an-option
reward up the top - that can be one pixel too high or low, but it's
close enough).

This makes the relevant code a lot cleaner, since it's not subtracting
stuff to try and figure out where it's space is for resource options.