Implement the boarder drone (but there's no combat yet)
It does board the enemy ship and seek out rooms with systems and crew,
but since crew combat isn't implemented yet it just sits there and does
nothing.

There were quite a few changes in various places to support it flying
out to the enemy ship, and making those visuals show (it counts as a
projectile).