Drive the crew animations from update
Instead of updateAnimation figuring out what the crewmember is currently
doing, set a variable in update to tell it. With more complicated
on-update logic (like the upcoming crew combat, this should avoid bugs
related to the animation and update logic differing, and the crew member
appears to do something other than what they actually are.