Fix resource boxes having misaligned contents
Previously we used a fixed width for each type of resource (fuel,
drones, missiles, scrap) in the resource box, but for longer resource
numbers (particularly double-digit negative ones) this wasn't enough.

This fixes it to be very similar to (but a few pixels off on the
right-hand side of the box, I can't figure out what sets the width
exactly in vanilla) the positioning and sizing in vanilla, so it looks
reasonable.