Fix rewards in multi-line options
Previously we used the length of the entire option to position the
resource box, so it went past the right-hand side of the window.

This fixes it to wrap the text earlier (depending on the width of the
resource box) and position it properly in Y if wrapping does occur.