Keep a list of crew in each room
Previously if we needed to check if a crewmember was in a room, this was
often done using the reserved slot system. This is a system designed for
pathfinding that we need to rewrite (it has to be kept in sync with
crew's location and existance, which is becoming an increasing source of
bugs), and isn't well suited for checking who is in a given room.

This replaces the reserved slots everywhere that isn't in some way
related to pathfinding, as those will be delt with later.