Recalculate the room pathfinding slots
Previously crew allocated and deallocated themselves to and from slots
as they walked about. This worked fine, except that as more and more
edge cases came in via teleporting it became incredibly fragile.

If crew were added or removed wrong they wouldn't take up the right
slots, and a slot in a room would suddenly become unavailable for other
crewmembers.

This fixes it by recalculating all of these slots based on the crew
positions and destinations whenever something changes. This means that
at worst, any inconsistency (caused by not recalculating the slots)
would be resolved when the next crewmember is moved around.