Add seeds for enemy ship generation
This now generates enemy ships from their own RNG. The seed of this RNG
is printed in the console, along with the name of the enemy ship
specification.

This can be used to debug the generation of an enemy ship. If it spawns
in some unexpected way, it's generation can be rerun while debugging by
using the new spawn-ship console command.

This was added because I found a ship spawning with two missiles (and
only two missiles) as offensive weapons, which shouldn't be allowed, but
wasn't able to debug it (I later found the seed: REBEL AAHrjH/f).