Shipgen: Fix the weapon type flags being ignored
Previously, we marked the ship as having a weapon satisfying one of the
two generation flags (is laser-like,  does hull damage) when we added it
to the list of weapon candidates, not when we actually picked it.

This meant you could get ships with two missiles, for example. This
patch fixes that.