Copy over OggInputStream from Slick
This has the same name, and as it should be earlier in the classpath
than Slick's version, we can modify it to make changes.

Notably there's a bug that prevents bp_laser_3.ogg from loading (related
to end-of-file handling) which we need to fix.

This patch copies in the original file, so our changes will be visible
in Git. If we committed the modified file at first, it wouldn't be
possible to get a nice diff view of the changes.