Fix loading events with no text crashing the game
When a dialogue window is opened, it loads it's first event. If that
first event doesn't have any text, the window is immediately closed.

The code in PlayerShipUI that opens the window ran it's constructor
before setting the open-window global, which is how it closed the
window. Thus the window wouldn't be closed, since the code to close it
would null out the open window variable before it was set.

This resulted in the window being rendered while it was in an invalid
state, crashing the game.