Implement scripted system power limits
This is the system that slugs use to cut off your oxygen supply or
medbay.

It supports all the modes of operation I could find, both to set and
clear these limits of players and enemies.

When you jump away from a beacon and come back, the effects are
re-imposed. I haven't checked if this is the case in vanilla FTL or not,
but it seems reasonable.