Fix teleporting out of an occupied room crashing
Previously, if one teleported a crewmember out of an enemy room, it
would (when the intruder->friendly mode changed) find if the crewmember
was standing on the same cell as one of the enemy ship's crew (which
they often would be, due to fighting), and command them to move to an
unoccupied cell.

This was done after all the movement stuff was cleared, and would thus
cause crew to walk through a door on the enemy ship after they'd
finished teleporting, causing a crash when targetDoor was used
(specifically, on line 267).