Don't translate rooms by the hull offset
Previously we drew the hull image at 0,0 and offset everything else to
match it. Since at some point we'll need to switch crew to using a real
pixel-accurate x,y position (rather than locking them to the grid, or
interpolating between cells), it's nice to get this done first.

This means there's a bit more of a clear 0,0 space for the ship, which
is why doing this first is attractive.