Make BoardingDrone use AbstractProjectile
This splits all the weapon-related stuff off into a new class,
AbstractWeaponProjectile, so we can use AbstractProjectile for the
boarding drone.

This cleans up BoardingDrone quite a bit, but more importantly will help
a lot with the pixel-based projectile work that will be added in an
upcoming patch. Rewriting it all for the boarding and (when implemented)
the hacking drone would be a very large pain, as it's significantly more
complicated.