Implement pixel-based projectile movement
Previously, projectiles had a set time taken to reach the enemy, and
that would determine it's position.

Now they have a position that's updated based on a pixel-per-second
speed, which matches how FTL works.

This lays the foundation for implementing combat and defense drones,
which depend on this kind of thing a lot more.