Fix the defence drone shooting at stuff it can't hit
This is where a projectile is going to hit it's target (note this
doesn't consider the shields to be a target), and the point at which the
defence drone's interceptor would reach it would be too late.

It's obviously a waste of a shot to fire at such targets, so this skips
them.