Implement the hacking system
This implements pretty much all the Hacking stuff that's not specific to
the different systems it can attack - those will all need to be
implemented separately.

Hacking probes are properly targeted by defence drones until they land,
and the turn-it-off-and-on-again trick to sneak a probe by a defence
drone works fine here too without any special handling (other than to
stop probes flying when powered off).