Add the reload-console command
This destroys the DebugConsole instance, so you can use Java's
hot-reloading system to add new commands without reloading the game.

Since we make the map of commands when the console instance is created,
we have to destroy it to make any entirely new commands appear
(modifying the on-run method of an existing command has always worked
fine).