Fix the hacking probe on enemy ships
Previously the hacking probe would fly to 1000px away from it's starting
point. For some Engi ships this meant it didn't get outside the -800 to
800px definition of a ship's space, and it thought it had reached it's
target.

This made it attempt to hack it's target room, which then immediately
removed the projectile as the hacking probe wasn't in the target ship's
projectiles list.