Write our own sound handling system
Previously we used Slick's wrapper around OpenAL, which wasn't
particularly flexible with stuff like pausing sounds.

We need that for looped sounds - these are things like the sound that
constantly plays while hacking (which was implemented for testing it),
which are paused when the player pauses the game.

This keeps the internal API for playing sounds mostly the same, but
looped sounds can now be used and they have a function (continueLoop)
that must be called every update or they stop playing. This is to avoid
looped sounds getting stuck playing over and over with no way to stop
them, which would be a very annoying kind of bug.