Fix text rendering never resetting the vertex count
Every frame it'd try to draw more and more vertices, instead of
resetting the number that should be drawn in the next draw call down to
zero at the end of the previous draw call.

I'm pretty amazed the game was working at all without it, though it
slowed the console down a lot.