Encapsulate a system's configuration for a ship
This moves all the information about a system that's loaded from a
ship's XML file (it's starting power, the computer position, etc) into a
new object.

This is an early preparation for loading/saving support: this lets us
separate out information that's never changed (like the computer
position) and load it separately from when the system object is created.

This touches every system class, as it removes the XML element that's
passed through all their constructors.