Split the ship loading and default equipment spawning
This means stuff that is always constant on a ship (such as it's room
and door positions, images, etc) are loaded when the Ship object is
constructed, but the default systems and equipment are loaded in a
separate method.

This will be very helpful for ship loading/saving, since we can then
only save and load stuff that can change, such as systems and equipment.