Rename SlickGame to InGameState
We'll want to implement stuff like menus which are completely separate
to the in-game stuff (eg, the main menu doesn't have any concept of a
player or enemy ship), which will mean making other states and selecting
which one to use.

It'll also help with testing saving/loading - if we can save and re-load
the game every update, that will make it much easier to find
serialisation bugs - and being able to clear everything out and load it
fresh will make this a lot easier.

Since SlickGame is very widely used, it's kept in a separate commit to
keep it separate from more reviewable changes, and to make it easier to
rebase stuff past this commit.