Implement initial saving and loading
This doesn't actually let you save the game and restart it. Instead, it
implements a way to save the game's state to XML so that'll be really
easy to implement.

Currently the only way to test it is via the dump-save and save-load
console commands. The former saves the game and prints the save out to
the console (it's just XML, so it's at least somewhat readable, if very
long), while save-load saves the game state then immediately loads it
back. The idea there is that if there's a bug in the serialisation or
deserialisation code, running save-load will make it's effect visible.

I also plan to add a debug flag that effectively runs save-load on every
update. Playing through the game with that on should shake out any
bugs, as anything that wasn't being saved would be immediately obvious.

Not everything is currently saved - notably weapons, drones and hacking.
Those will need to be implemented later.