Serialise the planet and background image at beacons
This uses a trick to avoid having to figure out a way of identifying the
background directly: a randomly-generated number is stored for each of
the background and planet, and that's used to determine which of the
possible images are shown.

This also means that if you re-visit a beacon, the images will be the
same as the first time you visited it.