Switch to a custotm FTLAnimation system
Previously we were using Slick's animation system, but it was basically
impossible to cleanly serialise (since it's based on an inaccessable
millisecond-to-next-frame timer).

This switches everything to a custom animation class, which stores the
time since the animation started as a float which can be easily
serialised.

This will be useful when implementing projectile serialisation in a
future patch.