Fix crew pathfinding crashing at high delta-times
Previously, when there was a very high delta-time for a frame (such as
that caused by pausing the game with a debugger), there was often a
crash at this line in AbstractCrew.updateMovement:

val current = pf.nodes.getValue(roomPos)

This happened because we checked if we got to our destination before
checking if we'd crossed into a new room. This is fine if we're running
slow enough to take at least two frames between entering a room and
reaching our destination, but at very high delta-times this was not
true.

The result was that findNearestRoomPos would produce an invalid
position, since it's not designed to deal with the player being outside
of the current room.