Fix crew slot conflict resolution
If two crewmembers do end up in the same slot, we have a system to send
one of them to a free slot to prevent the situation from persisting.

This was, unfortunately, broken - crew would check if they're already
standing in a cell before they check if it's occupied. This patch
reverses it, so they'll resolve these conflicts.

This can be tested by reverting the previous commit's change, and on the
Kestrel walking two crew into the medbay, then moving them both to the
shields room at the same time. This works quite a few - but not all -
rooms, so make sure you can reproduce it if you're testing something
like this.