Copy some basic UI stuff in cont-save-load mode
During continuous save and reload mode, all the UI stuff (for example,
what crew you've selected) is lost, making it practically impossible to
actually play in this mode.

This copies a few particularly important UI things over, in a way that's
very unlikely to compromise the purpose of the save-load test. This is
currently the crew selection rectangle and what crew are selected.

It also re-uses the same debug console, so it doesn't loose it's state
each update (the history was previously preserved, but the nice
interactive selection stuff wasn't).