Make the savegame a lot shorter
This makes the savegame take less likes when pretty-printed, in the
interests of making it easier to read for development purposes.

This mostly involves converting tags into attributes, or giving them a
default value and hiding them when that value is the current value (for
example, hiding the super-shield tag when the ship doesn't have a super
shield at all).

For the Kestrel loaded into a random sector, this reduces the length of
the savegame length from around 150 down to 100 lines.