Implement serialisation for hacking
This was surprisingly simple to implement.

The hacking probe is what does most of the work during deserialisation,
linking itself with the hacking system and telling the room under attack
that it's being hacked, rather than that deserialising a reference to
the hacking system (which ended up being a problem due to having to
register all the systems on all ships before being able to serialise
any of them, which looked fairly inconvenient).