Fix the crew dying animation resetting on loading
We used the dying animation to check if we should make the crew
disappear, but it's timer wasn't serialised. This meant the timer would
reset upon reloading the game.

This patch therefore saves the animation timer for all crew, regardless
of what they're doing. This also means the crew animation will play
properly when in continuous save/load mode.