Fix crew walking out of the ship after deserialisation
When being deserialised, the crew are first spawned like regular crew,
then load their XML.

Loading the XML didn't overwrite the current position they were walking
to if they were idle when saved, so when they just continued walking
towards whereever their last assigned point was. If there was a conflict
with what crew was in a cell, this could cause a crewmember to start
walking before they were deserialised.

This means they'd walk to their next waypoint, regardless of whether
there was a valid path or not.