Add 'save' and 'load' console commands
This lets you save the game, which is helpful if you're trying to debug
something that for whatever reason involves restarting the game, and is
at least somewhat tedious to set up.

These saves are stored in a directory inside the game's working
directory. This should probably be moved to a proper location, which we
can do when we add the proper (user-facing) save/load system.