Don't generate impossible sectors
When generating a sector, this ensures that there's a path from the
start to the exit beacon, and (matching FTL) tries to generate a sector
where that path is five or more jumps.

It also makes it possible for the exit beacon to occur in the
second-to-last column, which is unfortunately how it is in FTL too.