Make weapons not reference their weapons system
The only thing they really needed is the position they're supposed to
launch from, which is now set by the weapons system when the weapon is
bound to the weapons system.

This will be useful when implementing artillery, as we can do the same
thing for it to make it launch from the right place.