Fix the flagship crew getting stuck with artillery
Previously, we assumed that any crew could physically access any system.
That's obviously not the case with the flagship, so we need to avoid
assigning crew in the main body of the ship to the artillery rooms and
vice-versa.

This was causing crew in the body of the ship to get stuck and stand in
place, constantly failing to find a path to the artillery rooms.