Incorrectly implement artillery projectile weapons
This makes projectile weapons fire all their projectiles at a single
room, which is the exact opposite of the actual behaviour (firing
projectiles at different rooms). This is fairly complicated to
implement, and as such will be fixed in it's own patch.

This is at least enough to stop the flagship from crashing, however.

The exception is flak, which has different behaviour (it targets
separate rooms for different projectiles, but it's allowed for two
projectiles to target the same room).