Refactor the debug console to allow for custom output
Rather than the list of text lines in the console being a list of
strings, it's now a list of objects which have their draw function
called.

This will let us implement properly aligned text and stuff like that.

This changes every place in the console that writes a line of output,
hence this is split into a separate commit so later changes can be
meaningfully reviewed.