Line up the help message in the debug console
Along with the systems and flags list, this properly lines up all the
columns of the help message.

Previously we tried to by adding spaces to different lines, but this
works very badly with non-monospace fonts, and this system (lining them
up by drawing the strings in separate parts) looks a lot nicer.