Fix crew crashing during deserialisation
Previously, when calculating our room position, if we were pixel-aligned
but not grid-aligned we'd leave roomPosition as-is (it's supposed to be
null if the crew isn't grid-aligned).

This happened to work without serialisation, as the value would always
be nulled as the player would step to a non-pixel-aligned point at some
point before the value ended up pointing to the wrong room.

However, with cont-save-load mode on, walking a crewmember downwards on
an enmey ship would often cause a crash. This was from the roomPosition
being set incorrectly before the room was set, and then not being
updated later since the crew was pixel-aligned.