Correctly colour the enemy target box
Previously it was always it's neutral white colour, and never had the
red tint for hostile ships (which is how you normally see it).

This also makes a few small improvements to the text rendering, adding
the 'TARGET' text and using the non-legacy string drawing function,
along with localisation keys.