Add the GameText object to represent strings
This represents a string that may or may not be localised, which is what
all the blueprints are supposed to use.

This lets us properly load in non-localised names, which makes spawning
stuff in the console easier when there aren't the missing name strings
from cut (and thus not translated) items showing.