Remove the glow from drawIconAndPower x,y
In the AbstractSystem.drawIconAndPower method, we previously passed in
the coordinates the system icon image should be drawn at. This thus
included the 19 pixels of padding around it.

This was annoying to undo when drawing stuff in systems that override
that method, since you can't easily measure it out from screenshots.

Thus the coordinates are now at the corner of the non-transparent
region, if you drew a box that covers it. This matches the coordinates
passed into makeExtraButtons.