Rewrite the sound loop handling
Fires (which will be implemented in a subsequent patch) use it, and they
need to be able to handle multiple fires playing at the same time.

FTL does that by only playing one copy of the sound and varying it's
volume, so we need some clean way to do that.

The solution here is an extra layer of abstraction, a LoopHandle.
Anything that needs to play sound loops has a copy and calls a function
on each update when the sound is supposed to play.