Allow crew to turn around when walking through a door
Previously, if a crewmember was in a doorway when their target was
changed they'd always finish aligning themselves to the cell grid. This
was fine, except that if they were told to turn around and walk back
through the door this meant they'd walk all the way through before
turning around.

This patch does a bit of refactoring which isn't needed for this
specific change, but tells us what door the player is moving through.
This will be helpful when we implement lockable doors.