Draw crew portraits properly in event dialogue
This separates the information about a living crewmember (name, colour,
etc) from the actual crewmember instance, and uses that to render the
crew in the dialogue so they maintain their colour (and skills, once we
implement those) between the event window and when they're spawned.

This will also be very helpful for implementing crew in stores.