Properly render female humans
Obviously being added later in development, the 'human' sprites refer to
the male version, with a separate set of female sprites.

This sets up humans as having 8 'colours' (which are just a convenient
way of representing the appearance of the crew), with 4 variants of the
sprite for each gender.

This also picks names correctly, matching the gender between the crew's
name and sprite.