Switch to a custom Graphics implementation
The motivation behind this is to remove the high cost of popTransform,
allowing us to use it everywhere.

It also means that the bulk image rendering code can respect the current
transform, so you don't have to manually read the translation before
writing a string (this was done very little for performance reasons, but
it was an annoying limitation - particularly for the ship editor).

Finally, if we do want to move away from legacy OpenGL at some point,
this will make it a lot easier as we now have control over our graphics
class.