Fix the glow around drones
This is because the pixels around the drone are being drawn as
transparent white though it has a black edge, so the colour is averaged
out which makes the sprite glow.

They're rendered using nearest filtering, which actually does look ugly
once you look closely, but vanilla does it so it's fine to copy.