hangar: Block duplicate systems
This is particularly notable with the clonebay/medbay, which is a
special case.

We could allow the player to make a ship with the medbay/clonebay in
different rooms so they could both be purchased at once, but then we'd
need a way in the editor to have multiple systems per room. For this one
case, it really doesn't matter.

This is supposed to be an editor that sacrifices advanced capabilities
for ease-of-use, since Superluminal already exists.