Rewrite the crew AI task system to match FTL
This splits all the crew AI out into it's own class, and rewrites the
task system to match FTL's system, where tasks aren't allocated 1:1 with
crew.

This also means we have the AI object in place for all ships, so we can
use it for mind-controlled intruders on the player's ship.