ai: Fix boarders constantly switching tasks
This was an edge case with the system where boarders will prefer the
last one's task, since if that task was less important than multiple
other options, they'd constantly switch between those better options.

This patch fixes it by checking if the currently-assigned task is more
important than the task from the last crew, and if so ignoring that.