Fix engis rendering as humans
The colour outline would be correct for an engi, but the main image
would not.

The problem is that there's only a single engi colour, which is
represented internally by not having any colour filters. Thus the check
for if a crewmember is a female human (their colour ID being >= the
number of colours that crew has) broke.