Fix combat drones 'jolting' in an edge case
If you turn a combat drone off when it's slowing down to fire, it would
suddenly speed up and coast to a stop some distance from where it should
have.

This was because the power-off-slowdown and pause-to-fire-slowdown code
both used their own timers to calculate the current speed, and didn't
write it back to the speed variables. Fixing this properly would be a
pain, so now the pause-to-fire code is checked first as it can't be
toggled by the player, which should fix it.