hangar: Preserve the location of computers in rooms
When you first start editing a ship, the location of the computers in
each system are now copied into the EditableShip from the blueprint,
which prevents the computer positions from being broken when a ship is
switched to edit mode then played without making any changes.

This will also accomidate future changes to set the computer position in
the editor.