Reset the breach a crew is repairing when they move
This fixes a crash bug where a crew is repairing a breach in a 4x4 room,
and then walks into a 2x1 or 1x2 room. The slot index of the breach they
were repairing is now invalid.

Fire fighting deals with this by clearing that variable whenever the
crew is doing anything other than fighting a fire, so we can just copy
that for breach repairs.