Fix killed enemies always disappearing on save/load
This happens even if your crew are abord their ship, which is the one
case where we very much do not want the ship to just disappear.

This is solved by storing such ships with a special element, like the
flagship does, since they're no longer stored in a beacon.