Add the new UI system, use it for the game-over window
This adds an XML-based system for designing UIs. The motivation here is
that many windows are cut up into pieces that change size on the basis
of text width, which is a huge pain to get right using normal rendering.

Of particular note is the 'editor', which is just a viewer for these
XML-based UIs, and has stuff like zooming in.

This replaces the gameover window's bulky rendering code with an XML
file, which was vastly faster to write (and properly tiles the
background image).