Add support for multiple energy types
This means a system now tracks how much power it's drawing from each
potential source, whether it's the reactor, backup battery or (in the
future) Zoltans or some yet-to-be-created modded power type.

This is a fairly invasive change, and also moves the code for rendering
a system's power bars into that system itself, which opens up more
options for modding.