Fix Zoltans crashing the weapons system
When a Zoltan was powering the weapons system and it's power was then
decreased by the player, it'd repeatedly try to power itself down from
inside powerStateChanged. This would cause powerStateChanged to be
called again with the power increased back to the Zoltan-forced minimum,
leading to a StackOverflowError.

The mechanics of the weapons/zoltans interaction is still broken, but
now it's a behaviour mismatch vs vanilla not a crash.