console: Add blueprint autocompletion
This switches the 'weapon' and 'drone' commands to accepting the
blueprint name as an argument (rather than via a continuation), which
makes it easier to repeatedly spawn a weapon.

With this, blueprint-based completions are now implemented. For weapons
and drones, they're drawn as a button grid. Other blueprints (eg, crew
races) are drawn in a list.