ui: Add the 'free' container widget
Unlike hbox/vbox, this widget doesn't try to specially arrange it's
child elements.

It's intended to be used as a parent for widgets that are lined up over
images, where everything is specified with pixel positions.

An advantage of this (compared to making the contents of a UI a child of
it's images) is that you can subtract out the padding of the images, so
the positions of the widgets aren't relative to the padding.