Add profiles with achievements
Profiles are how we'll store stuff like high scores and achievements,
same as in vanilla.

There's currently no way to unlock these achievements in-game, but
there's a pair of debug commands (ach-set and ach-del) to manipulate
them for testing purposes.

These achievements are then used to show the locked/unlocked status of
the achievements in the pause menu.