Implement the music system
This should pretty well match vanilla, however there is currently one
major issue:

When switching between the combat and explore tracks, the entire OGG
file is decoded upto that point which is unfortunately quite slow,
blocking the main thread in that time.

This will be addressed in a subsequent commit.