Copy in jorbis's VorbisFile
This has a proper random seek implementation, which we'll use.

It has *quite* a few bugs from when it was converted from C (I'm
somewhat skeptical anyone actually ever used it), so we'll be making
some big modifications.

Vendoring it with a commit here will make the changes visible via a
future commit.