Fix the explore/combat music switching pause
This switches from using Slick's OGG parser to VorbisFile, which has
proper seeking support.

Unfortunately it was quite broken, so a bunch of fixes for it are
included.

The seeking time is now trivial, and the times match closely enough that
I couldn't hear them when rapidly reloading the explore track.