Fix fire and intruder damage showing damage numbers
When fire or intruders damage a system by one level, a popup number for
the one point of damage is currently generated.

This isn't supposed to happen, and what's worse is that when the system
is fully broken a second popup is generated for the hull damage.

This hides the former, while the latter (hull damage) popup is correct
and remains.