win: Detect and grab the ftl.dat path from FTL instances
This searches through the open windows for FTL, and finds the path of
ftl.dat from it.

This is useful for things like the Humble Bundle version of the game,
which aren't installed in a given location. It's easier to tell someone
to open vanilla up than to paste in the path, because experience suggets
there's a very good chance they'll make a mistake.