slipstream: Don't repeatedly parse big XML files
Previously, the Slipstream logic would convert files back and forth
between the in-memory XML representation and text for each mod that
modified the file.

This ended up taking about 90% of the time spent patching XML files, and
is obviously inefficient. Now all the XML is parsed once, and should
never be converted back to text again.