Implement caching for Slipstream XML patches
This stores the result of patches to vanilla XML files in a custom
binary format, which is very fast to load.

It doesn't store the result of parsing XML files not present in vanilla,
in the interests of reducing disc consumption (though this is probably
too paranoid, since it's only about 4MiB for Hyperspace and INS+). This
can be changed on Datafile.kt line 163.