Aggressively optimise the BXML parser
This uses sun.misc.Unsafe to skip a lot of JDOM's validation, in an
effort to parse in the XML as fast as possible.

It uses a custom input stream buffer, to avoid the synchronised read
method - fiddling with the object monitors is fast, but not if you're
doing it for every byte of integers you read.

The Element.setAttribute reimplementation only shows marginal gains, and
most of the remaining overhead is probably also not worth bothering
with.

Finally, it adds a JMH benchmark to test this.