Fix crew all teleporting to slot 0 at their target
We were assigning crew a slot on the new ship before changing the ship
they're currently on, so the mode variable wasn't properly updated.

Thus when teleporting to an enemy ship, the player's crew would check
for space for a CREW rather than an INTRUDER in the target room, and
they'd all find slot 0 was avaiable (assuming there wasn't already an
enemy there), and all teleport to it.