Store a single per-beacon seed
Previously we were storing a visual seed to set the background image,
instead store a single seed we'll use for everything.

In vanilla you can load an old save and all the beacons will spawn the
same ships, stores etc which is useful for 'what-if'ing a run if you
took save backups while playing it.

This is also useful for the automated tests, which obviously need to be
deterministic.