Make events deterministic
All the randomness in events now comes from a single seed set when the
event is first triggered, which for the events that show when you jump
to a new beacon is derived from the beacon's seed.

This essentially uses a Merkle tree of RNGs seeding each other as you go
down choices.

This also avoids storing localised event text in the savefile, since we
can just 'randomly' pick the same strings again. The same applies for
randomly-selected resources.