Fix the enemy not turning on weapons after empty missiles
If the enemy has a missile weapon that's out of ammo, they won't
atttempt to power up any subsequent weapon. This means that if their
first weapon is a missile, after running out of ammo they'll never turn
any of their weapons back on if their weapons system is damaged and
subsequently repaired.