hotkeys: Add a scroll bar
This doesn't support clicking to jump around, but it both indicates that
this is a scrollable area and shows the user's position in the area.

Previously there wasn't any cue that the area was scrollable, so a user
might think there were just a bunch of missing hotkeys for whatever
reason.