hotkey: Don't horizontally align groups
Previously we paired up groups, so if a large and small group were
paired together there'd be a lot of wasted space until the next pair of
groups.

We tried to avoid this by pairing groups of a similar size, but for
small groups of only a couple of items (such as the development/cheats
section, to be added in a later commit) didn't look great. It also lead
to a single group by itself at the top of the UI - the empty space on
the top-right looked a bit weird.

The two columns aren't locked together anymore, so there won't be any
wasted space.