Add a bunch of missing hotkeys
Previously there were quite a few more buttons on most keyboards that
you couldn't bind to hotkeys, since the game didn't know what to call
them.

This commit also fixes the menu key crashing the game - we were using
GLFW's last key constant as the number of avaiable keys, not the ID of
the last valid one. Thus we overran an array when it was pressed.