Consume fuel, add the out-of-fuel jump UI
It seems a little late to be adding fuel consumption after almost
everything else is done, but here it is.

This should more-or-less exactly match vanilla's out-of-fuel UI,
including the distress beacon sound (which was much more of a pain than
it sounds, since it's a sound loop that runs while the game is paused,
unlike all the others).