Abstract away the flagship logic
This moves the flagship logic into FlagshipBoss, and abstracts it behind
the BossManager interface.

This will let mods add custom bosses, and indeed have multiple bosses in
a single sector.

It'll make implementing the rest of the flagship logic a lot nicer,
since it's not stuffed into the Sector class anymore.

Finally, this also fixes the rebel shipyard event - previously that was
crashing when it loaded.